Move and refactor Name.java to battleship package
Fix checkstyle issues

Fix checkstyle issues
