+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -12,10 +12,10 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Cell;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
