+++ src/main/java/seedu/address/model/person/Cell.java
@@ -7,6 +7,7 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.battleship.Name;
 import seedu.address.model.tag.Tag;
 
 /**
