+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,10 +6,10 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.MapGrid;
 import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Cell;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
