+++ src/test/java/seedu/address/logic/commands/ListTagsCommandTest.java
@@ -16,7 +16,7 @@ import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Cell;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
