+++ src/test/java/seedu/address/storage/JsonAdaptedCellTest.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.testutil.Assert;
