+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -6,7 +6,7 @@ import java.util.Set;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Cell;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.util.SampleDataUtil;
