+++ src/test/java/seedu/address/ui/CellListPanelTest.java
@@ -20,7 +20,7 @@ import javafx.collections.ObservableList;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Cell;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.battleship.Name;
 import seedu.address.model.person.Phone;
 
 public class CellListPanelTest extends GuiUnitTest {
